requirements.txt
Flask==3.0.0
supabase==2.4.3
python-dotenv==1.0.0
gunicorn==21.2.0
